move rabbit.go into internal/metrics
