+++ internal/metrics/rabbit.go
